Fixes to pull request after @maximn review
 * Naming + comments
 * Better validation on exception type
 * File structure
