+++ hive-api-scala/src/main/scala/com/wix/hive/server/resolvers/springframework/package.scala
@@ -1,0 +1,7 @@
+package com.wix.hive.server.resolvers
+
+import org.springframework.web.context.request.WebRequest
+
+package object springframework {
+  type WixInstanceRequestResolver = PartialFunction[WebRequest, String]
+}
